requirements.txt
numpy==1.23.5
pandas==1.5.3
matplotlib==3.6.3
seaborn==0.12.2
streamlit
plotly
statsmodels
Pillow
